update_images/requirements.txt
scaleway-sdk==1.10.1